ingestion_function/requirements.txt
requests
google-cloud-secret-manager
google-cloud-storage
functions-framework
